6 user story

+++ db/migrate/20230923192648_create_discounts.rb
@@ -2,7 +2,7 @@ class CreateDiscounts < ActiveRecord::Migration[7.0]
   def change
     create_table :discounts do |t|
       t.integer :threshold
-      t.float :percentage
+      t.integer :percentage
       t.references :merchant, foreign_key: true
 
       t.timestamps
